% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himm-package.R
\docType{package}
\name{himm-package}
\alias{himm}
\alias{himm-package}
\title{himm: Hidden Integrated Markov Models in R and JAGS}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Authors:
\itemize{
  \item Maj Beldring Henningsen \email{majbh@sund.ku.dk}
}

}
\keyword{internal}
