% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_module.R
\name{load_module}
\alias{load_module}
\alias{unload_module}
\title{Load the internal JAGS module}
\usage{
load_module()

unload_module()
}
\description{
Load the internal JAGS module
}
\details{
This module provides the following distributions for JAGS:

LOMAX:  dlom(alpha, sigma)

\deqn{
  p(x) = \frac{\alpha}{\sigma} \left(1 + \frac{x}{\sigma}\right)^{-\left(\alpha+1\right)}
}{
  p(x) = (\alpha / \sigma) (1 + (x / \sigma)) ^ -(\alpha+1)
}

\deqn{\alpha > 0, \sigma > 0, x > 0}{\alpha > 0, \sigma > 0, x > 0}
}
