% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_hmm}
\alias{simulate_hmm}
\alias{simulate_basic}
\title{HMM data in one of two formats:  (1) simple matrix, (2) data frame}
\usage{
simulate_basic(
  N_animals = 100L,
  N_time = 10L,
  p1 = 0.1,
  beta_const = 0.05,
  beta_freq = 0.02,
  gamma = 0.08,
  sensitivity = 0.8,
  specificity = 0.99
)
}
\arguments{
\item{sp}{}
}
\description{
HMM data in one of two formats:  (1) simple matrix, (2) data frame
}
\examples{
Obs <- simulate_basic()
}
